Move all commands into cli/
