+++ tedge/src/cli/connect/az.rs
@@ -33,8 +33,8 @@ impl Azure {
             bridge_root_cert_path,
             remote_clientid: clientid,
             local_clientid: "Azure".into(),
-            bridge_certfile: config.query_string(DeviceCertPathSetting)?,
-            bridge_keyfile: config.query_string(DeviceKeyPathSetting)?,
+            bridge_certfile: config.query(DeviceCertPathSetting)?,
+            bridge_keyfile: config.query(DeviceKeyPathSetting)?,
             try_private: false,
             start_type: "automatic".into(),
             clean_session: true,
